This is our new version
To install, right click on the Roblox player icon in your desktop, on the menu click "Open file location", open the content folder AKA: "Content", and drag and drop the file to the bare folder called content. reopen your Roblox and join the game. You installed our custom content! [This is not a virus, its jus a folder with an image that loads in-game when installed, I recommend not to play on an account you have a lot of robux or work on because Roblox could detect it as a Unkown file, re install the file as stated every Roblox update to keep with the version.]